Refactor syncSender and syncScanner (3)
