+++ wire/conn_test.go
@@ -5,7 +5,9 @@ import (
 	"bytes"
 	"errors"
 	"io"
+	"net"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -124,4 +126,73 @@ func newEofReader(str string) io.ReadCloser {
 	limitReader := io.LimitReader(bytes.NewBufferString(str), int64(len(str)))
 	bufReader := bufio.NewReader(limitReader)
 	return io.NopCloser(bufReader)
+}
+
+func TestWriteMessage(t *testing.T) {
+	s, b := newTestSender()
+	err := s.SendMessage([]byte("hello"))
+	assert.NoError(t, err)
+	assert.Equal(t, "0005hello", b.String())
+}
+
+func TestWriteEmptyMessage(t *testing.T) {
+	s, b := newTestSender()
+	err := s.SendMessage([]byte(""))
+	assert.NoError(t, err)
+	assert.Equal(t, "0000", b.String())
+}
+
+func newTestSender() (Sender, *mockConn) {
+	w := new(mockConn)
+	return NewConn(w), w
+}
+
+// mockConn is a wrapper around a bytes.Buffer that implements io.Closer.
+type mockConn struct {
+	*bytes.Buffer
+}
+
+func makeMockConnStr(str string) net.Conn {
+	w := &mockConn{
+		Buffer: bytes.NewBufferString(str),
+	}
+	return w
+}
+
+func makeMockConnBuf(buf *bytes.Buffer) net.Conn {
+	w := &mockConn{
+		Buffer: buf,
+	}
+	return w
+}
+
+func makeMockConnBytes(b []byte) net.Conn {
+	w := &mockConn{
+		Buffer: bytes.NewBuffer(b),
+	}
+	return w
+}
+
+func (b *mockConn) Close() error {
+	// No-op.
+	return nil
+}
+
+func (b *mockConn) LocalAddr() net.Addr {
+	return nil
+}
+func (b *mockConn) RemoteAddr() net.Addr {
+	return nil
+}
+
+func (b *mockConn) SetDeadline(t time.Time) error {
+	return nil
+}
+
+func (b *mockConn) SetReadDeadline(t time.Time) error {
+	return nil
+}
+
+func (b *mockConn) SetWriteDeadline(t time.Time) error {
+	return nil
 }
