--- wire/sender_test.go
@@ -1,79 +1,0 @@
-package wire
-
-import (
-	"bytes"
-	"net"
-	"testing"
-	"time"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestWriteMessage(t *testing.T) {
-	s, b := NewTestSender()
-	err := s.SendMessage([]byte("hello"))
-	assert.NoError(t, err)
-	assert.Equal(t, "0005hello", b.String())
-}
-
-func TestWriteEmptyMessage(t *testing.T) {
-	s, b := NewTestSender()
-	err := s.SendMessage([]byte(""))
-	assert.NoError(t, err)
-	assert.Equal(t, "0000", b.String())
-}
-
-func NewTestSender() (Sender, *TestWriter) {
-	w := new(TestWriter)
-	return NewConn(w), w
-}
-
-// TestWriter is a wrapper around a bytes.Buffer that implements io.Closer.
-type TestWriter struct {
-	*bytes.Buffer
-}
-
-func MakeTestConn(str string) net.Conn {
-	w := &TestWriter{
-		Buffer: bytes.NewBufferString(str),
-	}
-	return w
-}
-
-func MakeTestConnBuf(buf *bytes.Buffer) net.Conn {
-	w := &TestWriter{
-		Buffer: buf,
-	}
-	return w
-}
-
-func MakeTestConnBytes(b []byte) net.Conn {
-	w := &TestWriter{
-		Buffer: bytes.NewBuffer(b),
-	}
-	return w
-}
-
-func (b *TestWriter) Close() error {
-	// No-op.
-	return nil
-}
-
-func (b *TestWriter) LocalAddr() net.Addr {
-	return nil
-}
-func (b *TestWriter) RemoteAddr() net.Addr {
-	return nil
-}
-
-func (b *TestWriter) SetDeadline(t time.Time) error {
-	return nil
-}
-
-func (b *TestWriter) SetReadDeadline(t time.Time) error {
-	return nil
-}
-
-func (b *TestWriter) SetWriteDeadline(t time.Time) error {
-	return nil
-}
