+++ wire/sync_conn.go
@@ -4,9 +4,54 @@ import (
 	"encoding/binary"
 	"fmt"
 	"io"
+	"net"
 	"os"
 	"time"
 )
+
+// SyncConn is a connection to the adb server in sync mode.
+// Assumes the connection has been put into sync mode (by sending "sync" in transport mode).
+// The adb sync protocol is defined at
+// https://android.googlesource.com/platform/system/core/+/master/adb/SYNC.TXT.
+// Unlike the normal adb protocol (implemented in Conn), the sync protocol is binary.
+// Lengths are binary-encoded (little-endian) instead of hex.
+// Length headers and other integers are encoded in little-endian, with 32 bits.
+// File mode seems to be encoded as POSIX file mode.
+// Modification time seems to be the Unix timestamp format, i.e. seconds since Epoch UTC.
+type ISyncConn interface {
+	io.Closer
+	StatusReader
+	ReadInt32() (int32, error)
+	ReadFileMode() (os.FileMode, error)
+	ReadTime() (time.Time, error)
+
+	// Reads an octet length, followed by length bytes.
+	ReadString() (string, error)
+
+	// Reads an octet length, and returns a reader that will read length
+	// bytes (see io.LimitReader). The returned reader should be fully
+	// read before reading anything off the Scanner again.
+	ReadBytes() (io.Reader, error)
+
+	// SendOctetString sends a 4-byte string.
+	SendOctetString(string) error
+	SendInt32(int32) error
+	SendFileMode(os.FileMode) error
+	SendTime(time.Time) error
+	// Sends len(data) as an octet, followed by the bytes.
+	// If data is bigger than SyncMaxChunkSize, it returns an assertion error.
+	SendBytes(data []byte) error
+}
+
+type SyncConn struct {
+	net.Conn
+	rbuf []byte
+	wbuf []byte
+}
+
+func NewSyncConn(r net.Conn) *SyncConn {
+	return &SyncConn{r, make([]byte, 4), make([]byte, 4)}
+}
 
 // ReadStatus reads a little-endian length from r, then reads length bytes and returns them
 func (s *SyncConn) ReadStatus(req string) (string, error) {
@@ -115,4 +160,53 @@ func readSyncMessage(r io.Reader, buf []byte) ([]byte, error) {
 		return buf[:n], fmt.Errorf("error reading message data: %w", err)
 	}
 	return buf[:n], nil
+}
+
+func (s *SyncConn) SendOctetString(str string) error {
+	if len(str) != 4 {
+		return fmt.Errorf("%w: octet string must be exactly 4 bytes: '%s'", ErrAssertion, str)
+	}
+
+	if n, err := s.Write([]byte(str)); err != nil {
+		return fmt.Errorf("error send string: %w, sent %d", err, n)
+	}
+	return nil
+}
+
+func (s *SyncConn) SendInt32(val int32) error {
+	if err := binary.Write(s, binary.LittleEndian, val); err != nil {
+		return fmt.Errorf("error sending int on sync sender: %w", err)
+	}
+	return nil
+}
+
+func (s *SyncConn) SendFileMode(mode os.FileMode) error {
+	if err := binary.Write(s, binary.LittleEndian, mode); err != nil {
+		return fmt.Errorf("error sending filemode on sync sender: %w", err)
+	}
+	return nil
+}
+
+func (s *SyncConn) SendTime(t time.Time) error {
+	if err := s.SendInt32(int32(t.Unix())); err != nil {
+		return fmt.Errorf("error sending time on sync sender: %w", err)
+	}
+	return nil
+}
+
+func (s *SyncConn) SendBytes(data []byte) error {
+	length := len(data)
+	if length > SyncMaxChunkSize {
+		// This limit might not apply to filenames, but it's big enough
+		// that I don't think it will be a problem.
+		return fmt.Errorf("%w: data must be <= %d in length", ErrAssertion, SyncMaxChunkSize)
+	}
+
+	if err := s.SendInt32(int32(length)); err != nil {
+		return fmt.Errorf("error sending data length on sync sender: %w", err)
+	}
+	if n, err := s.Write(data); err != nil {
+		return fmt.Errorf("error send bytes: %w, sent %d", err, n)
+	}
+	return nil
 }
