+++ wire/sync_conn_test.go
@@ -17,7 +17,7 @@ var (
 
 func TestSyncSendOctetString(t *testing.T) {
 	var buf bytes.Buffer
-	s := NewSyncConn(MakeTestConnBuf(&buf))
+	s := NewSyncConn(makeMockConnBuf(&buf))
 	err := s.SendOctetString("helo")
 	assert.NoError(t, err)
 	assert.Equal(t, "helo", buf.String())
@@ -25,13 +25,13 @@ func TestSyncSendOctetString(t *testing.T) {
 
 func TestSyncSendOctetStringTooLong(t *testing.T) {
 	var buf bytes.Buffer
-	s := NewSyncConn(MakeTestConnBuf(&buf))
+	s := NewSyncConn(makeMockConnBuf(&buf))
 	err := s.SendOctetString("hello")
 	assert.EqualError(t, err, "AssertionError: octet string must be exactly 4 bytes: 'hello'")
 }
 
 func TestSyncReadTime(t *testing.T) {
-	s := NewSyncConn(MakeTestConnBytes(someTimeEncoded))
+	s := NewSyncConn(makeMockConnBytes(someTimeEncoded))
 	decoded, err := s.ReadTime()
 	assert.NoError(t, err)
 	assert.Equal(t, someTime, decoded)
@@ -39,35 +39,35 @@ func TestSyncReadTime(t *testing.T) {
 
 func TestSyncSendTime(t *testing.T) {
 	var buf bytes.Buffer
-	s := NewSyncConn(MakeTestConnBuf(&buf))
+	s := NewSyncConn(makeMockConnBuf(&buf))
 	err := s.SendTime(someTime)
 	assert.NoError(t, err)
 	assert.Equal(t, someTimeEncoded, buf.Bytes())
 }
 
 func TestSyncReadString(t *testing.T) {
-	s := NewSyncConn(MakeTestConn("\005\000\000\000hello"))
+	s := NewSyncConn(makeMockConnStr("\005\000\000\000hello"))
 	str, err := s.ReadString()
 	assert.NoError(t, err)
 	assert.Equal(t, "hello", str)
 }
 
 func TestSyncReadStringTooShort(t *testing.T) {
-	s := NewSyncConn(MakeTestConn("\005\000\000\000h"))
+	s := NewSyncConn(makeMockConnStr("\005\000\000\000h"))
 	_, err := s.ReadString()
 	assert.Equal(t, errIncompleteMessage("bytes", 1, 5), err)
 }
 
 func TestSyncSendBytes(t *testing.T) {
 	var buf bytes.Buffer
-	s := NewSyncConn(MakeTestConnBuf(&buf))
+	s := NewSyncConn(makeMockConnBuf(&buf))
 	err := s.SendBytes([]byte("hello"))
 	assert.NoError(t, err)
 	assert.Equal(t, "\005\000\000\000hello", buf.String())
 }
 
 func TestSyncReadBytes(t *testing.T) {
-	s := NewSyncConn(MakeTestConn("\005\000\000\000helloworld"))
+	s := NewSyncConn(makeMockConnStr("\005\000\000\000helloworld"))
 
 	reader, err := s.ReadBytes()
 	assert.NoError(t, err)
